force lon_index and lat_index to be integers
I was getting this error:

Traceback (most recent call last):
  File "keweenawan_apw_path.py", line 368, in <module>
    plot_synthetic_paths(ax1, title='(a)')
  File "keweenawan_apw_path.py", line 139, in plot_synthetic_paths
    mcplates.plot.plot_distribution(myax, directions[:, 0],
directions[:, 1], cmap=next(dist_colors), resolution=60)
  File "/Users/Laurentia/0000_Github/mcplates/mcplates/plot.py", line
111, in plot_distribution
    lon_samples, lat_samples, resolution)
  File "/Users/Laurentia/0000_Github/mcplates/mcplates/plot.py", line
82, in density_distribution
    lon_grid, lat_grid, hist = bin_trace(lon_samples, lat_samples,
resolution)
  File "/Users/Laurentia/0000_Github/mcplates/mcplates/plot.py", line
73, in bin_trace
    hist[lat_index, lon_index] += 1
IndexError: only integers, slices (`:`), ellipsis (`...`),
numpy.newaxis (`None`) and integer or boolean arrays are valid indices
airbears2-10-142-152-72:keweenawan Laurentia$ python
keweenawan_apw_path.py 1 false 0,240,10

and changing these via int() seemed to be the most straight-forward fix

+++ mcplates/plot.py
@@ -44,7 +44,7 @@ continent_dictionary = { 'africa': 'af.asc',
 
 def bin_trace(lon_samples, lat_samples, resolution):
     """
-    Given a trace of samples in longitude and latitude, bin them 
+    Given a trace of samples in longitude and latitude, bin them
     in latitude and longitude, and normalize the bins so that
     the integral of probability density over the sphere is one.
 
@@ -68,9 +68,9 @@ def bin_trace(lon_samples, lat_samples, resolution):
             # Just skip invalid latitudes if they happen to arise
             continue
 
-        lon_index = np.floor((lon + 180.) / dlon)
-        lat_index = np.floor((lat + 90.) / dlat)
-        hist[lat_index, lon_index] += 1.
+        lon_index = int(np.floor((lon + 180.) / dlon))
+        lat_index = int(np.floor((lat + 90.) / dlat))
+        hist[lat_index, lon_index] += 1
 
     lat_grid += dlat / 2.
     lon_grid += dlon / 2.
@@ -130,7 +130,7 @@ def plot_distribution(ax, lon_samples, lat_samples, to_plot='d', resolution=30, 
 
 def plot_continent( ax, name, rotation_pole=None, angle=0.,  **kwargs):
     # Load the lat/lon file
-    datastream = pkgutil.get_data('mcplates', 'data/continents/' + continent_dictionary[name]) 
+    datastream = pkgutil.get_data('mcplates', 'data/continents/' + continent_dictionary[name])
     datalines = [line.strip()
                  for line in datastream.decode('ascii').split('\n') if line.strip()]
